JAX test for Laplacian calculation

+++ dreamplace/LaplacianCalc.py
@@ -8,7 +8,7 @@ import numpy as np
 def calc_laplacian(pin_list, edge_matrix):
     pin_id2idx: dict[int, int] = {pin_id: idx for idx, pin_id in enumerate(pin_list)}
     pin_count = pin_list.size
-    l_matrix = np.zeros((pin_count, pin_count), dtype=np.int32)
+    l_matrix = np.zeros((pin_count, pin_count), dtype=np.int16)
     for pins_in_a_net in edge_matrix:
         pins_of_interest = np.intersect1d(pins_in_a_net, pin_list)
         degree_plus_one = len(pins_of_interest)
